implementacion y tests load y subclases

+++ src/load/Reefer.java
@@ -8,7 +8,7 @@ package load;
  **/
 public class Reefer extends Load {
 	
-	private Double energyConsumption;
+	private int energyConsumption;
 	
 	/**
 	 * Constructor de la clase Reefer.
@@ -18,15 +18,13 @@ public class Reefer extends Load {
 	 * @param length La longitud de la carga en metros.
 	 * @param weight El peso de la carga en toneladas.
 	 * @param energyConsumption El consumo energetico en kw/hora.
-	 * 
-	 * @author Gabriela Fascetta
 	 */
-	public Reefer(Double width, Double height, Double length, Double weight, Double energyConsumption) {
+	public Reefer(Double width, Double height, Double length, Double weight, int energyConsumption) {
 		super(width, height, length, weight);
 		this.energyConsumption = energyConsumption;
 	}
 
 	public Double getEnergyConsumption() {
-		return energyConsumption;
+		return (double) energyConsumption;
 	}
 }
